Bug 624205: fix warning due to mismatched export declarations on MarkContext.
UltraBlame original commit: 558d826c33bfb472d96704e09ff12978f3199d29

+++ js/src/jstracer.h
@@ -1626,7 +1626,7 @@ class TraceRecorder
 
 
 
-            return tree->globalSlots->offsetOf(nativeGlobalSlot(vp)) == -1;
+            return tree->globalSlots->offsetOf((uint16)nativeGlobalSlot(vp)) == -1;
         }
         pendingGlobalSlotToSet = -1;
         return true;
